Init objects

+++ src/main/java/ru/mark/planets/models/Vector.java
@@ -1,27 +1,27 @@
 package ru.mark.planets.models;
 
 public class Vector {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public Vector(float x,float y) {
+    public Vector(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
